adds a dynamic network driver loader concept
This adds a dynamic loader to replace the currently static loading of
network device modules.  This allows NAPALM to load drivers dynamically
based on namespace

```
>>> import napalm
>>> driver = napalm.get_network_driver('napalm.eos')
>>> device = driver('veos01', 'admin', 'admin')
>>> device.open()
>>> device
<napalm.eos.EOSDriver instance at 0x10629d638>
>>> device.get_facts()
{'os_version': u'4.15.0F-2387143.4150F', 'uptime': 1309109, 'interface_list': [u'Ethernet1', u'Ethernet2', u'Ethernet3', u'Ethernet4', u'Ethernet5', u'Ethernet6', u'Ethernet7', u'Management1'], 'vendor': u'Arista', 'serial_number': u'', 'model': u'vEOS', 'hostname': u'veos01', 'fqdn': u'veos01.eng.ansible.com'}
```

Drivers do not have to be in the napalm namespace although for this to work.

```
>>> driver = napalm.get_network_driver('my.namespace.eos')
>>> device = driver('veos01', 'admin', 'admin')
>>> device.open()
>>> device
<my.namespace.eos.EOSDriver instance at 0x102193b68>
>>> device.get_facts()
```

This commit is just a proof of concept.  It includes on module (eos.py) that
has been modfied for testing purposes.