replaces static loader with dynamic loader
This commit replaces the static loading of drivers with a dynamic loader.  The
dynamic loader can load modules using either a full path namespace
or a single module name.  If the provided name is not a full path
namespace, then the module name is prepended with the default namespace.

For example, calling get_network_driver('foo') will be translated to
'napalm_foo.foo' but get_network_driver('napalm.foo') will not be
modified.