Junos < 13 does not have the table-name key
For this reason we need to build the request in such a way
that we are able to get the table counters even when that key is missing.
Under bgp-output-queue, there's a field called number that represents
the table ID, in the same oreder as the RIB tables under bgp-rib.
We can use this to get the right queue info, however this ID starts from 0
on older Junos and with 1 on newer versions.