Pass password when using ssh key
Currently it is not possible to use a ssh key with a passphrase.

`paramiko` defines `password` as follows:

```
password (str) – a password to use for authentication or for unlocking a
private key
```

This change passes the password when using a private key allowing the
use of a passphrase protected private key.