Add allow_agent optional argument support to junos
Situation here is a little different than with the netmiko-based drivers.

Junos driver uses python-eznc, which in turn uses ncclient for the netconf
parts, which then uses paramiko for the ssh transport.

In order to pass the allow_agent to paramiko, we have to pass it somehow to
python-eznc. Good thing it's already there, but gets activated only if
both password and private key arguments are None.

So, add an optional argument 'allow_agent' (default False), which,
when enabled, will call eznc's Device without password or private key,
effectively turning both of them to placeholders.