Raise SessionLockedException when there's already a session in progress
In the context of multiple configuration changes, there's a corner case when due
to a race condition, the device releases the config session (thus the session is
not listed under ``show configuration sessions``), however it didn't actually
finish committing. If there's another configuration change request in this exact
interval (i.e., between the moment we invoked ``commit_config`` and the moment
``commit_config`` finished executing), we will attempt to make a configuration
change and ``_lock`` in its current form won't detect it.
With this change, we check the ``config_session`` variable and in case NAPALM
didn't finish committing yet, would raise ``SessionLockedException``.
I run this change for a few months, and didn't notice any issues.