Cache BGP entries in get_route_to(longer=True)
When retrieving a large number of BGP routes, grab all the routes for
a specific VRF in a single go, instead of collecting each individual
route's BGP details.

Newer versions of EOS (Confirmed on 4.20.7M) do not allow simultaneous
use of the longer-prefixes and detail modifers for "show ip bgp", so
detect this and fall back, resulting in lost peer addresses for the routes.
Since this only occurs when `longer` is True, it should be OK until the
vendor addresses the behavior change from 4.19 to 4.20