Rework _find_txt implementation
Improve handling of YANG-modeled data. The new implementation now differentiates between a path match (returns a string) and a non-match (returns None unless a default is specified). In case a path matches a tag without a value, an empty string is returned. While this implementation is not truly YANG-aware, it facilitates the handling of any presence containers or empty leafs.