Use Kickstart version to determine NXOS version
`rr_sys_ver` reports the system version from before a reboot, not the current version. As the NXAPI seems to lack a straightforward `sys_ver_str` (This is why we’re in this mess in the first place), we’ll need to use the `kickstart_ver_str` instead.

This is also shown to be an issue in the official documentation: https://developer.cisco.com/docs/nx-api-cli-reference-for-the-cisco-nexus-9000-series-platform/#!version-commands/show-version

This fixes #1171