Add config filter for XR-only module set
Retrieval of configuration in XML format is limited to models in the XR-only module set. All the other models are filtered out. Both get_config and compare_config make use of this filtering. The RPC filter and namespace string added in constants.py. The models in the module set are determined using the YANG library model. If a device  doesn't support the YANG library model, no filtering is performed.