Use aply-groups inheritance on get-configuration RPC calls
The behaviour for get-configuration RPC calls has been changed in
2af90cc75, in order to ensure we're gathering from the right database.

The way junos-eznc works is that if there aren't any specific options
passed in, it uses apply-groups by default. But when there are options
passed in, as we do starting with 2af90cc75, it doesn't default the
values of `inherit` and `groups`, and therefore when gathering the
config it won't expand the apply-groups, see
https://github.com/Juniper/py-junos-eznc/blob/ab0e8603eac56fa1d049695c7ef86eb64582ab8f/lib/jnpr/junos/factory/cfgtable.py#L517-L520.

Changing this now, by adding a few new optional arguments, which allow
the user to select the desired behaviour.