Resolve #1181 and reduce the get_facts heavy workload
Primarily resolving #1181, switching to querying the MemorySummary from
the Operational block, rather than AdminOperational for XR 7+.

At the same time, I'm updating the RPC request from ``get_facts`` to
reduce the amount of data received from the device; from a VM it's
probably not a lot being received, but from an actual production device
there are walls of XML to be processed. This should therefore
significantly improve the facts gathering and ``get_environment`` too.