Move config lock check to _load_config
The config lock should always be checked whenever we create a new
configuration session, as we waste compute resources otherwise.  If two
NAPALM threads create a session, both will fail to commit with the
current order of operations -- moving the check to session creation
means the first one to obtain the lock will be able to commit.