eos: fix whitespace in test output for vrf
I believe when I first collected the output for this test case 6 years
ago, I didn't actually collect it from a router, and may have just
copied an existing text file and edited it manually.  This is an issue
since we are trying to switch to fixed-width column parsing for the
output.  I have hand-edited it this time to match the format and spacing
of the other test cases, as I do not have access to the router or config
that was used as the source for this test case.